updated to 1.18-pre5
