+++ build.gradle
@@ -35,7 +35,7 @@ dependencies {
 	// You may need to force-disable transitiveness on them.
 
 	// Cloth Config
-	modApi("me.shedaniel.cloth:cloth-config-fabric:6.0.39") {
+	modApi("me.shedaniel.cloth:cloth-config-fabric:6.0.42") {
 		exclude(group: "net.fabricmc.fabric-api")
 	}
 }
@@ -56,7 +56,7 @@ tasks.withType(JavaCompile).configureEach {
 	it.options.encoding = "UTF-8"
 
 	// Minecraft 1.17 (21w19a) upwards uses Java 16.
-	it.options.release = 16
+	it.options.release = 17
 }
 
 java {
