+++ gradle.properties
@@ -1,14 +1,14 @@
 # Done to increase the memory available to gradle.
 org.gradle.jvmargs=-Xmx1G
 
-minecraft_version=1.18-pre4
-yarn_mappings=1.18-pre4+build.5
+minecraft_version=1.18-pre5
+yarn_mappings=1.18-pre5+build.2
 loader_version=0.12.5
 
 #Fabric api
-fabric_version=0.42.7+1.18
+fabric_version=0.42.8+1.18
 
 # Mod Properties
-	mod_version = 1.1.12
+	mod_version = 1.1.13
 	maven_group = com.kwpugh
-	archives_base_name = MobCatcher-Fabric-1.18-pre4
+	archives_base_name = MobCatcher-Fabric-1.18-pre5
