+++ src/main/resources/mob_catcher.mixins.json
@@ -2,7 +2,7 @@
   "required": true,
   "minVersion": "0.8",
   "package": "com.kwpugh.mob_catcher.mixin",
-  "compatibilityLevel": "JAVA_17",
+  "compatibilityLevel": "JAVA_16",
   "mixins": [
     "AbstractDonkeyEntityMixinInteract",
     "VillagerEntityMixinInteract",
